requirements.txt
crewai[google-genai]
crewai-tools
duckduckgo-search
requests
python-dotenv
langchain-google-genai
langchain-community
litellm
google-genai
yfinance
